Avoid class-string in ApiResources :(
Unfortunately, we can't write this code in a typesafe way and also have
the unit tests work, because the tests are running in this codebase,
where nothing in the \App namespace exists at all. Either we fail to get
a string back from the method at all (if we assert(class_exists())
inside it), or PHPStan correctly points out that the test is trying to
compare a string that does not refer to an existing class name with a
class-string.

So we've just gotta fall back on returning regular strings, sadly.
Sighs.