Use sleight of hand for type-compat with old Carbon
Older versions of Carbon::createFromTimestamp() declare that the
timestamp needs to be an integer, but a numeric-string will actually
work just fine. In newer Carbon, numeric-strings are handled differently
and will support bigger numbers, which we would like to take advantage
of.

We could use a cast to int to placate PHPStan, but that would mean we
lose the ability to work with really big timestamps ourselves. This
approach is compatible with both the oldeer and newer Carbon packages
and will also be approved by PHPStan (because we're basically tricking
it).