Swap PHPUnit printer so we can lock Orchestra versions
Since this package is a dependency for others, it kinda needs to try to
work with different versions of its dependencies. To that end, in the
Github workflow we install its deps with both --prefer-stable and
--prefer-lowest, which helps test that at least two versions of the deps
work okay.

Unfortunately that meant pulling in a completely different major version
of orchestra/testbench, which was a problem because the older one just
crashed PHPUnit entirely. I guess PHPUnit itself changed?

I couldn't insist on getting version 7 of the Orchestral packages
because they require a newer Symfony YAML package, and the result
printer we were using required older versions of that YAML package. No
idea why it needs a YAML parser to pretty-print PHPUnit results, but
whatever.

This new result printer package doesn't depend on anything and shouldn't
cause further dependency hell. It also looks a little nicer in my
opinion. I like BDD.